requirements.txt
asn1crypto==1.5.1
certifi==2024.7.4
cffi==1.15.1
charset-normalizer==3.3.2
coincurve==18.0.0
idna==3.7
numpy==1.21.6
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
Pillow==9.5.0
pycparser==2.21
pycryptodome==3.20.0
requests==2.31.0
torch==1.13.1
torchvision==0.14.1
typing-extensions==4.7.1
urllib3==2.0.7
